Add evidence box photo documentation
Basically all the evidence left over from the ceremony was packaged up - in
particular the EMI shielded box and remains of the compute node laptop - and
sealed in a 18in³ box. The seams of the box were taped with thread for tamper
evidentness; I forgot to buy glitter nail polish and had to catch a flight
before I could buy it. That box was then in turn put in a 24in³ box for storage
and potential later shipment.